#347 put getAgeString tests in the correct file
